# LMNN: Bounds for avoiding Impostors recalculation