--- a/src/mlpack/methods/lmnn/constraints.hpp
+++ b/src/mlpack/methods/lmnn/constraints.hpp
@@ -56,7 +56,8 @@ class Constraints
    */
   void TargetNeighbors(arma::Mat<size_t>& outputMatrix,
                        const arma::mat& dataset,
-                       const arma::Row<size_t>& labels);
+                       const arma::Row<size_t>& labels,
+                       const arma::vec& norms);
 
   /**
    * Calculates k similar labeled nearest neighbors for a batch of dataset and
@@ -71,6 +72,7 @@ class Constraints
   void TargetNeighbors(arma::Mat<size_t>& outputMatrix,
                        const arma::mat& dataset,
                        const arma::Row<size_t>& labels,
+                       const arma::vec& norms,
                        const size_t begin,
                        const size_t batchSize);
 
@@ -84,7 +86,8 @@ class Constraints
    */
   void Impostors(arma::Mat<size_t>& outputMatrix,
                  const arma::mat& dataset,
-                 const arma::Row<size_t>& labels);
+                 const arma::Row<size_t>& labels,
+                 const arma::vec& norms);
 
   /**
    * Calculates k differently labeled nearest neighbors & distances to
@@ -98,7 +101,8 @@ class Constraints
   void Impostors(arma::Mat<size_t>& outputNeighbors,
                  arma::mat& outputDistance,
                  const arma::mat& dataset,
-                 const arma::Row<size_t>& labels);
+                 const arma::Row<size_t>& labels,
+                 const arma::vec& norms);
 
   /**
    * Calculates k differently labeled nearest neighbors for a batch of dataset
@@ -113,6 +117,7 @@ class Constraints
   void Impostors(arma::Mat<size_t>& outputMatrix,
                  const arma::mat& dataset,
                  const arma::Row<size_t>& labels,
+                 const arma::vec& norms,
                  const size_t begin,
                  const size_t batchSize);
 
@@ -131,6 +136,7 @@ class Constraints
                  arma::mat& outputDistance,
                  const arma::mat& dataset,
                  const arma::Row<size_t>& labels,
+                 const arma::vec& norms,
                  const size_t begin,
                  const size_t batchSize);
 
@@ -150,6 +156,7 @@ class Constraints
                  arma::mat& outputDistance,
                  const arma::mat& dataset,
                  const arma::Row<size_t>& labels,
+                 const arma::vec& norms,
                  const arma::uvec& points,
                  const size_t numPoints);
 
@@ -163,7 +170,8 @@ class Constraints
    */
   void Triplets(arma::Mat<size_t>& outputMatrix,
                 const arma::mat& dataset,
-                const arma::Row<size_t>& labels);
+                const arma::Row<size_t>& labels,
+                const arma::vec& norms);
 
   //! Get the number of target neighbors (k).
   const size_t& K() const { return k; }
@@ -196,6 +204,14 @@ class Constraints
   * and different datapoints on the basis of labels.
   */
   inline void Precalculate(const arma::Row<size_t>& labels);
+
+  /**
+  * Re-order neighbors on the basis of increasing norm in case
+  * of ties among distances.
+  */
+  inline void ReorderResults(const arma::mat& distances,
+                             arma::Mat<size_t>& neighbors,
+                             const arma::vec& norms);
 };
 
 } // namespace lmnn