--- a/src/mlpack/methods/lmnn/constraints_impl.hpp
+++ b/src/mlpack/methods/lmnn/constraints_impl.hpp
@@ -37,11 +37,52 @@ Constraints<MetricType>::Constraints(
   }
 }
 
+template<typename MetricType>
+inline void Constraints<MetricType>::ReorderResults(
+                                            const arma::mat& distances,
+                                            arma::Mat<size_t>& neighbors,
+                                            const arma::vec& norms)
+{
+  // Shortcut...
+  if (neighbors.n_rows == 1)
+    return;
+
+  // Just a simple loop over the results---we want to make sure that the
+  // largest-norm point with identical distance has the last location.
+  for (size_t i = 0; i < neighbors.n_cols; i++)
+  {
+    for (size_t start = 0; start < neighbors.n_rows - 1; start++)
+    {
+      size_t end = start + 1;
+      while (distances(start, i) == distances(end, i) &&
+          end < neighbors.n_rows)
+      {
+        end++;
+        if (end == neighbors.n_rows)
+          break;
+      }
+
+      if (start != end)
+      {
+        // We must sort these elements by norm.
+        arma::Col<size_t> newNeighbors =
+            neighbors.col(i).subvec(start, end - 1);
+        arma::uvec indices = arma::conv_to<arma::uvec>::from(newNeighbors);
+
+        arma::uvec order = arma::sort_index(norms.elem(indices));
+        neighbors.col(i).subvec(start, end - 1) =
+            newNeighbors.elem(order);
+      }
+    }
+  }
+}
+
 // Calculates k similar labeled nearest neighbors.
 template<typename MetricType>
 void Constraints<MetricType>::TargetNeighbors(arma::Mat<size_t>& outputMatrix,
                                               const arma::mat& dataset,
-                                              const arma::Row<size_t>& labels)
+                                              const arma::Row<size_t>& labels,
+                                              const arma::vec& norms)
 {
   // Perform pre-calculation. If neccesary.
   Precalculate(labels);
@@ -59,6 +100,10 @@ void Constraints<MetricType>::TargetNeighbors(arma::Mat<size_t>& outputMatrix,
     knn.Train(dataset.cols(indexSame[i]));
     knn.Search(k, neighbors, distances);
 
+    // Re-order neighbors on the basis of increasing norm in case
+    // of ties among distances.
+    ReorderResults(distances, neighbors, norms);
+
     // Re-map neighbors to their index.
     for (size_t j = 0; j < neighbors.n_elem; j++)
       neighbors(j) = indexSame[i].at(neighbors(j));
@@ -74,6 +119,7 @@ template<typename MetricType>
 void Constraints<MetricType>::TargetNeighbors(arma::Mat<size_t>& outputMatrix,
                                               const arma::mat& dataset,
                                               const arma::Row<size_t>& labels,
+                                              const arma::vec& norms,
                                               const size_t begin,
                                               const size_t batchSize)
 {
@@ -102,6 +148,10 @@ void Constraints<MetricType>::TargetNeighbors(arma::Mat<size_t>& outputMatrix,
     knn.Train(dataset.cols(indexSame[i]));
     knn.Search(subDataset.cols(subIndexSame), k, neighbors, distances);
 
+    // Re-order neighbors on the basis of increasing norm in case
+    // of ties among distances.
+    ReorderResults(distances, neighbors, norms);
+
     // Re-map neighbors to their index.
     for (size_t j = 0; j < neighbors.n_elem; j++)
       neighbors(j) = indexSame[i].at(neighbors(j));
@@ -115,7 +165,8 @@ void Constraints<MetricType>::TargetNeighbors(arma::Mat<size_t>& outputMatrix,
 template<typename MetricType>
 void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputMatrix,
                                         const arma::mat& dataset,
-                                        const arma::Row<size_t>& labels)
+                                        const arma::Row<size_t>& labels,
+                                        const arma::vec& norms)
 {
   // Perform pre-calculation. If neccesary.
   Precalculate(labels);
@@ -132,6 +183,10 @@ void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputMatrix,
     // set and  same class points as query set.
     knn.Train(dataset.cols(indexDiff[i]));
     knn.Search(dataset.cols(indexSame[i]), k, neighbors, distances);
+
+    // Re-order neighbors on the basis of increasing norm in case
+    // of ties among distances.
+    ReorderResults(distances, neighbors, norms);
 
     // Re-map neighbors to their index.
     for (size_t j = 0; j < neighbors.n_elem; j++)
@@ -148,7 +203,8 @@ template<typename MetricType>
 void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputNeighbors,
                                         arma::mat& outputDistance,
                                         const arma::mat& dataset,
-                                        const arma::Row<size_t>& labels)
+                                        const arma::Row<size_t>& labels,
+                                        const arma::vec& norms)
 {
   // Perform pre-calculation. If neccesary.
   Precalculate(labels);
@@ -165,6 +221,10 @@ void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputNeighbors,
     // set and  same class points as query set.
     knn.Train(dataset.cols(indexDiff[i]));
     knn.Search(dataset.cols(indexSame[i]), k, neighbors, distances);
+
+    // Re-order neighbors on the basis of increasing norm in case
+    // of ties among distances.
+    ReorderResults(distances, neighbors, norms);
 
     // Re-map neighbors to their index.
     for (size_t j = 0; j < neighbors.n_elem; j++)
@@ -182,6 +242,7 @@ template<typename MetricType>
 void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputMatrix,
                                         const arma::mat& dataset,
                                         const arma::Row<size_t>& labels,
+                                        const arma::vec& norms,
                                         const size_t begin,
                                         const size_t batchSize)
 {
@@ -209,6 +270,10 @@ void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputMatrix,
     // set and same class points as query set.
     knn.Train(dataset.cols(indexDiff[i]));
     knn.Search(subDataset.cols(subIndexSame), k, neighbors, distances);
+
+    // Re-order neighbors on the basis of increasing norm in case
+    // of ties among distances.
+    ReorderResults(distances, neighbors, norms);
 
     // Re-map neighbors to their index.
     for (size_t j = 0; j < neighbors.n_elem; j++)
@@ -226,6 +291,7 @@ void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputNeighbors,
                                         arma::mat& outputDistance,
                                         const arma::mat& dataset,
                                         const arma::Row<size_t>& labels,
+                                        const arma::vec& norms,
                                         const size_t begin,
                                         const size_t batchSize)
 {
@@ -253,6 +319,10 @@ void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputNeighbors,
     // set and same class points as query set.
     knn.Train(dataset.cols(indexDiff[i]));
     knn.Search(subDataset.cols(subIndexSame), k, neighbors, distances);
+
+    // Re-order neighbors on the basis of increasing norm in case
+    // of ties among distances.
+    ReorderResults(distances, neighbors, norms);
 
     // Re-map neighbors to their index.
     for (size_t j = 0; j < neighbors.n_elem; j++)
@@ -271,6 +341,7 @@ void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputNeighbors,
                                         arma::mat& outputDistance,
                                         const arma::mat& dataset,
                                         const arma::Row<size_t>& labels,
+                                        const arma::vec& norms,
                                         const arma::uvec& points,
                                         const size_t numPoints)
 {
@@ -298,6 +369,10 @@ void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputNeighbors,
     knn.Search(dataset.cols(points.elem(subIndexSame)),
         k, neighbors, distances);
 
+    // Re-order neighbors on the basis of increasing norm in case
+    // of ties among distances.
+    ReorderResults(distances, neighbors, norms);
+
     // Re-map neighbors to their index.
     for (size_t j = 0; j < neighbors.n_elem; j++)
       neighbors(j) = indexDiff[i].at(neighbors(j));
@@ -313,18 +388,19 @@ void Constraints<MetricType>::Impostors(arma::Mat<size_t>& outputNeighbors,
 template<typename MetricType>
 void Constraints<MetricType>::Triplets(arma::Mat<size_t>& outputMatrix,
                                        const arma::mat& dataset,
-                                       const arma::Row<size_t>& labels)
+                                       const arma::Row<size_t>& labels,
+                                       const arma::vec& norms)
 {
   // Perform pre-calculation. If neccesary.
   Precalculate(labels);
 
   size_t N = dataset.n_cols;
 
-  arma::Mat<size_t> impostors;
-  Impostors(impostors, dataset);
+  arma::Mat<size_t> impostors(k, dataset.n_cols);
+  Impostors(impostors, dataset, labels, norms);
 
-  arma::Mat<size_t> targetNeighbors;
-  TargetNeighbors(targetNeighbors, dataset);
+  arma::Mat<size_t> targetNeighbors(k, dataset.n_cols);;
+  TargetNeighbors(targetNeighbors, dataset, labels, norms);
 
   outputMatrix = arma::Mat<size_t>(3, k * k * N , arma::fill::zeros);