see #1448