@rcurtin 2018-07-10
Ah, I see also that this (like others) has a potential problem with the storage of the old transformation matrix for different batches.  Let's talk about that specifically in #1461 so we (or more specifically I) don't get confused during the discussion because it is in multiple places. :)