@manish7294 2018-07-11
Here are some simulations. https://gist.github.com/manish7294/b1efd856ba3013803be29801bfa4629e
I have used mean-centering here. Have a doubt about the correctness of it tough:
```
for (size_t i = 0; i < data.n_rows; i++)
  {
    data.row(i) =  data.row(i) - (double)arma::mean(data.row(i));
  }
```
It seems points change too much on covertype.