@rcurtin 2018-07-12
I believe that's correct, looking at the implementation.  If you really want to quickly check if it is correct, just ensure that the starting kNN accuracy with and without mean-centering is the same (kNN should be invariant to mean-centering).  The real check would be if the exact kNN results including distances are the same, so you could do that too.

The results look great.  I can see that in the later iterations for vc2 and iris, very few impostors are recalculated.  I'm not too surprised about the covertype dataset---I've found that the points have very large norm and so as a result the bounds don't seem to be as effective there.  Still the mean centering should help.

When #1461 is done let's return to this and work out the batch size issue, do some more timing simulations to convince ourselves the speedup is good, and then merge it in. :+1: