@manish7294 2018-07-13
Just have to figure out one issue - we are obliged to calculate `k + 1` target neighbors and impostors instead of k. Now, let say we have only 6 points in one of the class. Ideally, we should be able to run simulations up to k = 5, but now we are restricted to up to k = 4. That's the reason one of the unit test is failing. 
I think If we see this problem from the point of real datasets, we generally don't have this low number of points in any class and shouldn't cause much of a problem. We may just output an error message to the user if this is happening with his/her parameters. What do you think?