@rcurtin 2018-07-13
Right, that won't typically happen in real-world datasets, but if it does you can do this... search for `k` neighbors, resize the distances matrix to have `k + 1` rows, then fill that last row with `DBL_MAX`.  Then the bounds will still work correctly.