@manish7294 2018-07-13
Ya, that would solve the problem, just an extra check like this inside LMNNFunction() will be required

```
  size_t minCount = arma::min(arma::histc(labels, arma::unique(labels)));
  if (minCount < k + 1)
  {
    // use DBL_MAX
  }
```