@rcurtin 2018-07-13
Couldn't you just check the size of each of the training sets right before the `Search()` call instead of calling `histc()`?