@manish7294 2018-07-20
@rcurtin It looks like when we calculate `targetNeighbors` for `k + 1`, then the result of first k rows is a bit different than when we calculate `targetNeighbors` for `k`. Ideally, the first k rows of result from `k + 1` targetNeighbors calculation should be the same as the result when `k` targetNeighbors are calculated. Similarly, `Impostors` faces the same issue.

To cross-check this thing, I ran `mlpack_knn` twice, one for `k = 3` and one for ` k = 4`, and it seems entries of some of the columns gets a kind of shuffling / re-ordering effect on changing `k`.