@rcurtin 2018-07-20
Can you file an issue for the KNN bug so that I can try and reproduce it?  If the distances are exactly the same, the results may come back in a slightly different order, but the distance should always be exactly the same.