@manish7294 2018-07-20
One of the Accuracy tests over 6 points simple dataset using amsgrad is failing somehow, and it's weird that I am not getting that on my system. Debbuging this led me #1470 earlier, as there is some deviation in final results because of that. Other than that, I don't think I am still up to anything else.