@manish7294 2018-07-20
I performed some simulations and I think having a dataset having different distances to all neighbors solves the issue. Even vc2 works perfectly.