@manish7294 2018-07-20
I think the deviation could be expected as we are actually training over a different permutation of points. Here a point may be such that it has a greater impact over some other dimension than the other point, and still having same distance from the query point.