@rcurtin 2018-07-20
I see.  Give me some time to think about it, and you can think about it also, and let's see if we can figure out a way to handle it.  The answer may be requiring that NearestNeighborSearch gives us back deterministic orderings of points, or the answer may be that we specifically look through our kNN results and reorder neighbor indices when distances are the same.