@manish7294 2018-07-20
On a second thought, Though being a bit different, the distance learned will still be correct, as we are just taking a different set of constraints which satisfy the conditions of LMNN.