@rcurtin 2018-07-21
Unfortunately I've done a little bit of thinking about the exact-same-distance issue and I've come to two conclusions:

1.  Right now all bounds we are computing are with respect to the k'th impostor.  In essence these bounds are saying, "the distance between the point and its k'th impostor next iteration cannot be greater than (our bound quantity)".  But remember that this depends on the norm of the impostor.  So what if the (k - 1)'th impostor has huge norm but the k'th impostor has very small norm?  In this case it is possible that the (k - 1)'th impostor will be further away next iteration than the k'th impostor.  That means that, with respect to our bounds, it's possible that our current bounds could say "it's not possible that the (k + 1)'th impostor and the k'th impostor could switch places in the distance order" and be correct---but at the exact same time it could be possible that the (k - 1)'th and (k + 1)'th impostor would switch places.  (Or something like this.)  At the end of my long explanation here, it's clear that the only solution is that we have to rework the bounds a little, to take the max over all `k` possible bounds.  (Also note that for `k == 1`, there is no problem, and I think the bounds being incorrect would be very rare in practice.)  Let's worry about this issue in another PR or something since I don't believe it is causing any of our issues.

2.  The ordering of the results does matter because it will affect the gradient direction.  If the impostor ordering is switching every iteration, this will cause trouble for the gradient (I believe).  In any case, the order in which we want our results to be is in order of increasing norm.  So, e.g., if the k'th and (k + 1)'th impostor have the same distance to the query point, then we'd prefer to have it in the order that the k'th impostor has smaller norm than the (k + 1)'th impostor.  But since this is problem-specific to our LMNN optimizations, it doesn't make sense to me to do this computation in the `NeighborSearch` code.  So I believe the best way to do this will be to update `Impostors()` so that it runs a loop after calling `Search()` to ensure that the ordering is correct.  I worked out what I think is some code that will do this, but I did not have a chance to fully integrate this into the different `Constraints::Impostors()` overloads.  On the laptop I'm using I don't actually have a way to middle-click, so I can't copy-paste it, but I'll email the function to you separately.