@rcurtin 2018-07-28
Just an update, I have not collated my results entirely yet but I have seen promising results on large covertype datasets and MNIST:

```
covertype-5k lbfgs: 66 iterations

        total     computing_neighbors     acc before      acc after
old:    80.758s   21.672s                 75.56%          79.98%
new:    75.456s   15.992s                 75.56%          79.98%

covertype-5k amsgrad: 50 passes, batch size 256

        total     computing_neighbors     acc before      acc after
old:    136.093s  16.157s                 75.56%          75.54%
new:    137.484s  16.813s                 75.56%          77.26%


        total     computing_neighbors     building    acc before      acc after
old:    249.279s  128.153s                5.587s      90.11%          91.796%
new:    620.516s  340.776s                ~100s       90.11%          91.862%
```

The first MNIST run, on the 10k test points (784d) with L-BFGS and 25 maximum iterations, took ~3670s before this optimization and ~3697s afterwards.  So I guess that there is not much optimization possible here, but I want to investigate that more.  Alternately, if you would like to determine whether or not any impostors are being pruned with the MNIST dataset I'd be interested to see the results.

Overall I'm fairly convinced this is good and we should merge it, but I just want to make sure and see a few more results (mostly with large datasets) first.  I will be busy this weekend though, so I won't really have a chance to run anything, so if you're able to get to it before me great, otherwise feel free to wait.